Allows moving a share into another, (filesharing)
When a share is moved inside another, will check if both shares have equivalence in sharing types, permissions and sharing targets. If they have, will remove the sharing of the first and move it inside the second.

This changes a policy in nextcloud and I don't know if this will be accepted.
https://github.com/nextcloud/server/issues/6707

Signed-off-by: Vinicius Cubas Brand <viniciuscb@gmail.com>